m5stack_core::driver::radio::wifi

Enum AuthenticationMethod

#[non_exhaustive]
pub enum AuthenticationMethod {

    None,
    Wep,
    Wpa,
    Wpa2Personal,
    WpaWpa2Personal,
    Wpa2Enterprise,
    Wpa3Personal,
    Wpa2Wpa3Personal,
    WapiPersonal,
    Owe,
    Wpa3EntSuiteB192Bit,
    Wpa3ExtPsk,
    Wpa3ExtPskMixed,
    Dpp,
    Wpa3Enterprise,
    Wpa2Wpa3Enterprise,
    WpaEnterprise,

}

Supported Wi-Fi authentication methods.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

None

No authentication (open network).

Wep

Wired Equivalent Privacy (WEP) authentication.

Wpa

Wi-Fi Protected Access (WPA) authentication.

Wpa2Personal

Wi-Fi Protected Access 2 (WPA2) Personal authentication (default).

WpaWpa2Personal

WPA/WPA2 Personal authentication (supports both).

Wpa2Enterprise

WPA2 Enterprise authentication.

Wpa3Personal

WPA3 Personal authentication.

Wpa2Wpa3Personal

WPA2/WPA3 Personal authentication (supports both).

WapiPersonal

WLAN Authentication and Privacy Infrastructure (WAPI).

Owe

Opportunistic Wireless Encryption (OWE)

Wpa3EntSuiteB192Bit

WPA3 Enterprise Suite B 192-bit Encryption

Wpa3ExtPsk

This authentication mode will yield same result as AuthenticationMethod::Wpa3Personal and is not recommended to be used. It will be deprecated in future, please use AuthenticationMethod::Wpa3Personal instead.

Wpa3ExtPskMixed

This authentication mode will yield same result as AuthenticationMethod::Wpa3Personal and is not recommended to be used. It will be deprecated in future, please use AuthenticationMethod::Wpa3Personal instead.

Dpp

Wi-Fi DPP / Wi-Fi Easy Connect

Wpa3Enterprise

WPA3-Enterprise Only Mode

Wpa2Wpa3Enterprise

WPA3-Enterprise Transition Mode

WpaEnterprise

WPA-Enterprise security